backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
openai==1.3.8
supabase==2.0.3
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
stripe==7.9.0
pillow==10.1.0
httpx==0.25.2
python-dotenv==1.0.0
alembic==1.13.0